+++ admin/users.php
@@ -1,6 +1,7 @@
 <?php
 require_once '../includes/functions.php';
 require_once '../includes/header.php';
+requireRole("admin");
 global $db;
 $users = $db->fetchAll("SELECT * FROM users WHERE role !='admin' ORDER BY first_name");
 ?>
@@ -8,23 +9,25 @@ $users = $db->fetchAll("SELECT * FROM users WHERE role !='admin' ORDER BY first_
 <div class="admin-container">
 
 
-  <?php foreach ($users as $user) { ?>
-    <div class="admin-item">
-        <form action="" method="POST">
-            <div class="admin-item-body">
-                <h5 class="card-title"><?php echo $user['first_name'] . " " . $user['last_name']?></h5>
-                <p class="card-text"><?php echo $user['username']?></p>
-                <p class="card-text"><?php echo $user['role']?></p>
-                <div class="admin-buttons">
-                    <a href="./user.php?request=edit" class="btn btn-success btn-lg">Edit</a>
-                    <a href="./user.php?request=delete" class="btn btn-danger btn-lg">Delete</a>
+    <?php if (empty($users)) { ?>
+        <p>No users found.</p>
+    <?php } else { ?>
+        <?php foreach ($users as $user) { ?>
+            <div class="admin-item">
+                <div class="admin-item-body">
+                    <h5 class="card-title"><?php echo htmlspecialchars($user['first_name'] . " " . $user['last_name']); ?></h5>
+                    <p class="card-text">Username: <?php echo htmlspecialchars($user['username']); ?></p>
+                    <p class="card-text">Role: <?php echo htmlspecialchars($user['role']); ?></p>
+                    <div class="admin-buttons">
+                        <a href="./user.php?request=edit&id=<?php echo urlencode($user['id']); ?>" class="btn btn-success btn-lg">Edit</a>
+                        <a href="./user.php?request=delete&id=<?php echo urlencode($user['id']); ?>" class="btn btn-danger btn-lg" onclick="return confirm('Are you sure you want to delete this user?');">Delete</a>
+                    </div>
                 </div>
             </div>
-        </form>
+        <?php } ?>
+    <?php } ?>
+</div>
 
-    </div>
-    <?php }?>
-
-    <?php
-    require_once '../includes/footer.php';
-    ?>
+<?php
+require_once '../includes/footer.php';
+?>
